+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Add
 + Add Axios
 + Add Axios interceptors
++ Add Router
++ Add navigation
 
 ## 0.2.1