+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-# 0.2.1
+# 0.3.0
+
+### Add
++ Add Axios
++ Add Axios interceptors
++ Add Router
++ Add navigation
+
+## 0.2.1
 
 ### Fix
 + Fix change data from API